requirements.txt
Django==1.7
PyYAML==3.11
